Fix and one more feature
The added closing of Self Service(+) didn't have a subtitle causing it to always fail - fixed
The added restart if Migration Assistant hasn't been started, now only appears if there are no failed installs.
And added the option to have swiftDialog open in dark or light themes. Default matches the macOS default of the light theme.